Har fikset populate klassen. Har indsat test data. Har kørt http fil igennem for at tjekke at vi får dataen ud. Alt kører. Har lavet en test klasse til EventDAO klassen.

+++ src/main/java/dat/daos/impl/EventGroupDAO.java
@@ -77,7 +77,7 @@ public class EventGroupDAO implements IDAO<EventGroupDTO, Integer> {
             EventGroup r = em.find(EventGroup.class, integer);
             r.setEventGroupNumber(eventGroupDTO.getEventGroupNumber());
             r.setEventGroupType(eventGroupDTO.getEventGroupType());
-            r.setEventGroupPrice(BigDecimal.valueOf(eventGroupDTO.getEventGroupPrice()));
+            r.setEventGroupPrice(eventGroupDTO.getEventGroupPrice());
             EventGroup mergedEventGroup = em.merge(r);
             em.getTransaction().commit();
             return new EventGroupDTO(mergedEventGroup);
